generator/pyproject.toml
[tool.poetry]
name = "nba stats api generator"
version = "1.0.0"
description = "Generates OpenAPI Json files to document stats.nba.com"
authors = ["Abhi Agarwal <abhiaagarwal01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
